+++ maps/include/maps/G3SkyMap.h
@@ -155,27 +155,27 @@ public:
 	virtual void ApplyMask(const G3SkyMapMask &mask, bool inverse=false);
 
 	// Pointing information
-	std::vector<size_t> AnglesToPixels(const std::vector<double> & alphas,
+	std::vector<ssize_t> AnglesToPixels(const std::vector<double> & alphas,
 	    const std::vector<double> & deltas) const;
 	void PixelsToAngles(const std::vector<size_t> & pixels,
 	    std::vector<double> & alphas, std::vector<double> & deltas) const;
-	std::vector<size_t> QuatsToPixels(const G3VectorQuat &quats) const;
+	std::vector<ssize_t> QuatsToPixels(const G3VectorQuat &quats) const;
 	G3VectorQuat PixelsToQuats(const std::vector<size_t> &pixels) const;
 
 	virtual std::vector<double> PixelToAngle(size_t pixel) const = 0;
-	virtual size_t AngleToPixel(double alpha, double delta) const = 0;
+	virtual ssize_t AngleToPixel(double alpha, double delta) const = 0;
 	quat PixelToQuat(size_t pixel) const;
-	size_t QuatToPixel(quat q) const;
+	ssize_t QuatToPixel(quat q) const;
 
 	// Rebinning and interpolation
-	virtual void GetRebinAngles(long pixel, size_t scale,
+	virtual void GetRebinAngles(size_t pixel, size_t scale,
 	    std::vector<double> & alphas, std::vector<double> & deltas) const = 0;
-	G3VectorQuat GetRebinQuats(long pixel, size_t scale) const;
+	G3VectorQuat GetRebinQuats(size_t pixel, size_t scale) const;
 	virtual void GetInterpPixelsWeights(double alpha, double delta,
-	    std::vector<long> & pixels, std::vector<double> & weights) const = 0;
-	void GetInterpPixelsWeights(quat q, std::vector<long> & pixels,
+	    std::vector<ssize_t> & pixels, std::vector<double> & weights) const = 0;
+	void GetInterpPixelsWeights(quat q, std::vector<ssize_t> & pixels,
 	    std::vector<double> & weights) const;
-	double GetInterpPrecalc(const std::vector<long> &pixels,
+	double GetInterpPrecalc(const std::vector<ssize_t> &pixels,
 	    const std::vector<double> &weights) const;
 	double GetInterpValue(double alpha, double delta) const;
 	double GetInterpValue(quat q) const;
@@ -186,10 +186,10 @@ public:
 	virtual boost::shared_ptr<G3SkyMap> Rebin(size_t scale, bool norm = true) const = 0;
 
 	/* Analogue to healpy.query_disc, returns list of pixels within a disc */
-	virtual std::vector<long> QueryDisc(double alpha, double delta, double radius) const {
+	virtual std::vector<ssize_t> QueryDisc(double alpha, double delta, double radius) const {
 		throw std::runtime_error("QueryDisc not implemented");
 	};
-	std::vector<long> QueryAlphaEllipse(double alpha, double delta, double a, double b) const;
+	std::vector<ssize_t> QueryAlphaEllipse(double alpha, double delta, double a, double b) const;
 
 	virtual bool IsDense() const {
 		throw std::runtime_error("Checking array density not implemented");
